fix: login guest user
